Move sb types to shared folder
